requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
networkx>=2.6.0
scipy>=1.7.0
scikit-learn>=1.0.0

# LLM APIs
openai>=1.0.0
anthropic>=0.3.0
google-generativeai>=0.3.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.62.0
plotly>=5.0.0
kaleido>=0.2.1

# Data analysis
statsmodels>=0.13.0
pingouin>=0.5.0

# Development
pytest>=6.2.0
black>=21.0.0
flake8>=4.0.0
